rf: split and empty utils module of interfaces
TSV2JSON|DictMerge|JoinTSVColumns|AddTSVHeader -> to `utility` module
TPM2ROI|AddTPMs -> to `probmaps` module
